+++ client/src/components/Upload/components/controllers/index.js
@@ -1,1 +1,7 @@
-export {default as UploadFilenameController} from './upload-filename.controller';
+export {default as uploadFilenameController} from './upload-filename.controller';
+export {default as trackTitleController} from './track-title.controller';
+export {default as artistController} from './artist.controller';
+export {default as bpmController} from './bpm.controller';
+export {default as genreController} from './genre.controller';
+export {default as trackDurationController} from './track-duration.controller';
+export {default as trackKeyController} from './track-key.controller';
